refactor: 기타 패키지 리팩터링
