+++ src/main/java/com/nhnacademy/frontend/payment/controller/PaymentController.java
@@ -1,8 +1,8 @@
 package com.nhnacademy.frontend.payment.controller;
 
-import com.nhnacademy.frontend.payment.domain.request.PaymentApprovalRequestDto;
-import com.nhnacademy.frontend.payment.domain.request.PaymentRequestDto;
-import com.nhnacademy.frontend.payment.domain.response.PaymentResponseDto;
+import com.nhnacademy.frontend.payment.dto.request.PaymentApprovalRequestDto;
+import com.nhnacademy.frontend.payment.dto.request.PaymentRequestDto;
+import com.nhnacademy.frontend.payment.dto.response.PaymentResponseDto;
 import com.nhnacademy.frontend.payment.service.PaymentService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
